benchmarks/requirements.txt
torch>=2.2
torchvision>=0.17
timm>=1.0.0
numpy>=1.24
pillow>=10.0
pyyaml>=6.0
wandb>=0.17
tqdm>=4.66

fvcore>=0.1.5.post20221221
thop>=0.1.1.post2209072238

matplotlib>=3.8

lpips>=0.1.4
